Removed the Tokio dependent binary.
