+++ .markdownlint.jsonc
@@ -1,0 +1,3 @@
+{
+  "MD013": { "code_blocks": false }
+}
