+++ ls_dcm/Cargo.toml
@@ -13,10 +13,8 @@ dicom-core.workspace = true
 dicom-object.workspace = true
 dicom-dictionary-std.workspace = true
 thiserror.workspace = true
-anyhow.workspace = true
 clap.workspace = true
 walkdir.workspace = true
-tokio.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 comfy-table.workspace = true
@@ -25,7 +23,3 @@ rayon.workspace = true
 [[bin]]
 name = "ls_rtplan"
 path = "src/bin/ls_rtplan.rs"
-
-[[bin]]
-name = "ls_rtplan_r"
-path = "src/bin/ls_rtplan_r.rs"
