+++ ls_dcm/src/bin/ls_rtplan.rs
@@ -1,15 +1,16 @@
-use std::path::{Path, PathBuf};
+use rad_tools_ls_dcm::model::{DicomFile, Modality};
+use rad_tools_ls_dcm::DicomError;
+use rayon::prelude::*;
+use std::ffi::OsString;
+use std::path::PathBuf;
 use std::time::SystemTime;
 
 use clap::Parser;
-use comfy_table::modifiers::UTF8_ROUND_CORNERS;
-use comfy_table::presets::UTF8_FULL;
-use comfy_table::Table;
-use tracing::{debug, error, Level, trace, warn};
-use walkdir::{DirEntry, WalkDir};
+use tracing::{debug, error, trace, warn, Level};
+use walkdir::WalkDir;
 
 use rad_tools_ls_dcm::io::read_dicom_file_partial_by_modalities;
-use rad_tools_ls_dcm::model::{DicomFile, Modality, SopClass};
+use rad_tools_ls_dcm::view;
 
 /// A command line interface (CLI) application for reading and listing RTPLAN DICOM files.
 ///
@@ -17,13 +18,18 @@ use rad_tools_ls_dcm::model::{DicomFile, Modality, SopClass};
 /// as well as additional options such as filtering by filename prefixes, limiting the number of displayed results,
 /// sorting the files by last modified timestamp, and enabling logging at different levels.
 #[derive(Parser, Debug, Clone)]
-#[command(author, version, about, long_about = "
+#[command(
+    author,
+    version,
+    about,
+    long_about = "
 A command line interface (CLI) application for reading and listing RTPLAN DICOM files.
 
 Application enables the user to specify the directory from which the DICOM files are read,
 as well as additional options such as filtering by filename prefixes, limiting the number of displayed results,
 sorting the files by last modified timestamp, and enabling logging at different levels.
-")]
+"
+)]
 struct Cli {
     /// Directory from where DICOM files are read (recursively).
     /// If unspecified, the current directory (".") is analysed.
@@ -42,16 +48,26 @@ struct Cli {
     /// Enable logging at DEBUG level.
     #[arg(long, default_value_t = false)]
     debug: bool,
-    /// Only use one thread to list the (DICOM) files in the directory.
-    #[arg(long, default_value_t = false)]
-    single_threaded_listing: bool,
     /// Enable logging at TRACE level.
     #[arg(long, default_value_t = false)]
     trace: bool,
 }
 
-#[tokio::main]
-async fn main() {
+#[derive(thiserror::Error, Debug)]
+pub enum AppError {
+    #[error("Error transversing the directory / file tree.")]
+    WalkDir(#[from] walkdir::Error),
+    #[error("Directory / file entry filtered out: {0:#?}")]
+    FilteredOut(OsString),
+    #[error("Unable to obtain last modified timestamp from {0:#?}")]
+    ModifiedTimestamp(OsString),
+    #[error("IO error while transvering directory tree.")]
+    IO(#[from] std::io::Error),
+    #[error(transparent)]
+    DicomError(#[from] DicomError),
+}
+
+fn main() {
     let mut cli = Cli::parse();
     let level = if cli.trace {
         Level::TRACE
@@ -73,294 +89,101 @@ async fn main() {
     }
 
     let dir_path = PathBuf::from(cli.dir.as_ref().unwrap());
-    let mut files = if !cli.single_threaded_listing {
-        ls_files2(dir_path, &cli.prefixes, |entry, prefixes| -> bool {
-            if !entry.path().is_file() {
-                return false;
-            }
-            return if !prefixes.is_empty() {
-                entry
-                    .file_name()
-                    .to_str()
-                    .map(|s| prefixes.iter().any(|prefix| s.starts_with(prefix)))
-                    .unwrap_or(false)
-            } else {
-                true
-            };
-        },
-        ).await
-    } else {
-        ls_files(dir_path, |entry| -> bool {
-            if !entry.path().is_file() {
-                return false;
-            }
-            return if !cli.prefixes.is_empty() {
-                entry
-                    .file_name()
-                    .to_str()
-                    .map(|s| cli.prefixes.iter().any(|prefix| s.starts_with(prefix)))
-                    .unwrap_or(false)
-            } else {
-                true
-            };
-        })
+
+    let filter = |entry: &walkdir::DirEntry, prefixes: &[String]| -> bool {
+        if !entry.path().is_file() {
+            return false;
+        }
+        return if !prefixes.is_empty() {
+            entry
+                .file_name()
+                .to_str()
+                .map(|s| prefixes.iter().any(|prefix| s.starts_with(prefix)))
+                .unwrap_or(false)
+        } else {
+            true
+        };
     };
-    trace!("List of filtered files:");
-    for (_t, entry) in &files {
-        trace!("Path: {:#?}", entry.path());
-    }
-    if cli.sort {
-        trace!("Sorting DICOM files by last modified timestamp (reversed).");
-        files.sort_by(|(t1, _), (t2, _)| t1.cmp(t2));
-        files.reverse();
-    }
-    let files2: Vec<_> = files.into_iter().map(|(_, entry)| entry).collect();
     let modalities = vec![Modality::RtPlan];
-    let dicom_files = read_dicom_files(files2, modalities, cli.limit).await;
-    let view_model_items = build_view_model(&dicom_files);
-    let view = build_view(&view_model_items);
+
+    trace!("Starting to iterate the directory: {:#?}", &dir_path);
+    let par_iter = WalkDir::new(dir_path)
+        .into_iter()
+        .par_bridge()
+        .map(|r| read_dicom_data(r, filter, &cli.prefixes, &modalities));
+    let results: Vec<Result<EntryData, AppError>> = par_iter.collect();
+
+    trace!("Filtering out warnings and errors.");
+    let mut dataset = vec![];
+    for r in results {
+        if r.as_ref().is_err() {
+            debug!("{:#?}", r);
+            continue;
+        }
+        dataset.push(r.unwrap());
+    }
+
+    if cli.sort {
+        trace!("Sorting DICOM data by last modified timestamp (reversed).");
+        dataset.sort_by(|a, b| a.modified_time.cmp(&b.modified_time));
+        dataset.reverse();
+    }
+
+    let has_limit = cli.limit.is_some();
+    let limit = cli.limit.unwrap_or_default();
+    let mut dicom_files = vec![];
+    for data in dataset {
+        dicom_files.push(data.dicom_file);
+        if has_limit && dicom_files.len() >= limit {
+            break;
+        }
+    }
+
+    let view_model = view::build_model(&dicom_files);
+    let view = view::build_view(&view_model);
     println!("{view}");
 }
 
-fn build_view(items: &[ViewModelItem]) -> Table {
-    trace!("Building tabluar view.");
-    let hdr = ["Patient ID", "Patient Name", "Plan Name", "Plan Label"];
-
-    let mut table = Table::new();
-    table.load_preset(UTF8_FULL);
-    table.apply_modifier(UTF8_ROUND_CORNERS);
-    table.set_header(hdr);
-    for item in items {
-        table.add_row([&item.patient_id, &item.patient_name, &item.plan_name, &item.plan_label]);
-    }
-    table
+#[derive(Clone, Debug)]
+struct EntryData {
+    modified_time: SystemTime,
+    dicom_file: DicomFile,
 }
 
-/// Represents a view model item that contains information about a DICOM file.
-#[derive(Clone, Debug, Default)]
-struct ViewModelItem {
-    patient_id: String,
-    patient_name: String,
-    sop: SopClass,
-    plan_name: String,
-    plan_label: String,
-    dose_grids: usize,
-}
-
-fn build_view_model(dfs: &[DicomFile]) -> Vec<ViewModelItem> {
-    let mut view = vec![];
-    for df in dfs {
-        if let DicomFile::RTPlan(plan) = df {
-            view.push(ViewModelItem {
-                patient_id: plan.patient_id.clone(),
-                patient_name: plan.patient_name.clone(),
-                sop: plan.sop.clone(),
-                plan_name: plan.plan_name.clone(),
-                plan_label: plan.plan_label.clone(),
-                dose_grids: 0,
-            });
-        }
-    }
-    for df in dfs {
-        if let DicomFile::RTDose(dose) = df {
-            for rsc in &dose.referenced_rtplan_sequence {
-                for v in &mut view {
-                    if rsc.ref_class_uid == v.sop.class_uid
-                        && rsc.ref_instance_uid == v.sop.instance_uid
-                    {
-                        v.dose_grids += 1;
-                    }
-                }
-            }
-        }
-    }
-    view
-}
-
-/// Reads DICOM files asynchronously based on the provided list of files and modalities.
-///
-/// # Arguments
-///
-/// * `files` - A vector of `DirEntry` representing the files to be read.
-/// * `modalities` - A vector of `Modality` representing the modalities to filter the DICOM data.
-/// * `limit` - An optional usize representing the limit of DICOM files to read.
-///
-/// # Returns
-///
-/// A vector of `DicomFile` representing the DICOM files read.
-async fn read_dicom_files(
-    files: Vec<DirEntry>,
-    modalities: Vec<Modality>,
-    limit: Option<usize>,
-) -> Vec<DicomFile> {
-    let mut handles = vec![];
-    debug!("Starting tasks to read DICOM files");
-    for file in &files {
-        let t_file = file.clone();
-        let t_modalities = modalities.clone();
-        let handle = tokio::spawn(async move {
-            read_dicom_file_partial_by_modalities(t_file.path().to_path_buf(), t_modalities)
-        });
-        handles.push(handle)
-    }
-    debug!("Waiting for all reading tasks to complete");
-    let mut dicom_files = vec![];
-    for handle in handles {
-        match &limit {
-            None => {}
-            Some(limit) => {
-                if dicom_files.len() >= *limit {
-                    handle.abort();
-                    continue;
-                }
-            }
-        }
-
-        let result_join = handle.await;
-        if result_join.as_ref().is_err() {
-            trace!("Unable to read DICOM data: {:#?}", result_join.err());
-            continue;
-        }
-        if let Ok(v) = result_join.unwrap() {
-            dicom_files.push(v);
-        }
-    }
-    debug!("All reading tasks completed.");
-    dicom_files
-}
-
-/// Returns a list of files in the specified directory that pass the given filter function.
-/// Each file is represented by a tuple containing its last modified timestamp and `DirEntry` struct.
-///
-/// # Arguments
-///
-/// * `path` - A path to the directory.
-/// * `filter` - A closure that takes a `DirEntry` as parameter and returns a boolean value indicating whether the file should be included in the result
-#[allow(dead_code)]
-fn ls_files<P: AsRef<Path>, F>(path: P, filter: F) -> Vec<(SystemTime, DirEntry)>
-    where
-        F: Fn(&DirEntry) -> bool,
+fn read_dicom_data<F>(
+    r: Result<walkdir::DirEntry, walkdir::Error>,
+    filter: F,
+    prefixes: &[String],
+    modalities: &[Modality],
+) -> Result<EntryData, AppError>
+where
+    F: Fn(&walkdir::DirEntry, &[String]) -> bool,
 {
-    let path = path.as_ref();
-    debug!("Listing files in: {:#?}", path);
-    let mut v = vec![];
-    for entry in WalkDir::new(path).into_iter().flatten() {
-        if filter(&entry) {
-            if let Ok(meta) = entry.metadata() {
-                if let Ok(modified) = meta.modified() {
-                    v.push((modified, entry));
-                } else {
-                    warn!("Unable to obtain last modified timestamp from {:#?}", entry);
-                }
-            } else {
-                warn!("Unable to obtain meta data from {:#?}", entry);
-            }
-        }
+    let entry = r?;
+    if !filter(&entry, prefixes) {
+        return Err(AppError::FilteredOut(dir_entry_path(&entry)));
     }
-    debug!("Listing files in: {:#?} completed", path);
-    v
+    let metadata = entry.metadata()?;
+    let modified_time = metadata.modified().map_err(|e| {
+        let filename = dir_entry_path(&entry);
+        warn!(
+            "Unable to get last modified time of {:#?} due to an error: {:#?}",
+            &filename, e
+        );
+        AppError::ModifiedTimestamp(filename)
+    })?;
+    let dicom_file = read_dicom_file_partial_by_modalities(entry.path(), modalities)?;
+
+    Ok(EntryData {
+        modified_time,
+        dicom_file,
+    })
 }
 
-
-#[derive(thiserror::Error, Debug)]
-pub enum FileListError {
-    #[error("Unable to iterate {0:#?}")]
-    IterationError(PathBuf),
-    #[error("Unable to obtain last modified timestamp from {0:#?}")]
-    ModifiedTimestamp(PathBuf),
-    #[error("Unable to obtain meta data from {0:#?}")]
-    MetaData(String),
-    #[error("File path not retained after filter: {0:#?}")]
-    FilteredOut(PathBuf),
-}
-
-/// Returns a list of files in the specified directory that pass the given filter function.
-/// Each file is represented by a tuple containing its last modified timestamp and `DirEntry` struct.
-///
-/// # Arguments
-///
-/// * `path` - A path to the directory.
-/// * `filter` - A closure that takes a `DirEntry` as parameter and returns a boolean value indicating whether the file should be included in the result
-async fn ls_files2<P, F>(path: P, prefixes: &Vec<String>, filter: F) -> Vec<(SystemTime, DirEntry)>
-    where P: AsRef<Path> + Send + Sync + 'static,
-          F: (FnOnce(DirEntry, Vec<String>) -> bool) + Copy + Send + Sync + 'static
-{
-    let path = path.as_ref().to_path_buf();
-    debug!("Listing files in: {:#?}", path);
-
-    let tasks = WalkDir::new(path.clone()).into_iter().map(|entry| {
-        let entry = match entry {
-            Ok(entry) => entry,
-            Err(err) => {
-                error!("{:#?}", err);
-                return Err(FileListError::IterationError(path.clone()));
-            }
-        };
-
-        let tpath = path.clone();
-        let tentry = entry.clone();
-        let tprefixes = prefixes.clone();
-        let task = tokio::task::spawn(async move {
-            let t = tentry.clone();
-            if filter(t.clone(), tprefixes) {
-                let result_meta = t.metadata();
-                match result_meta {
-                    Ok(meta) => {
-                        match meta.modified() {
-                            Ok(modified_time) => { Ok((modified_time, t)) }
-                            Err(e) => {
-                                error!("{:#?}", e);
-                                Err(FileListError::ModifiedTimestamp(tpath))
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        error!("{:#?}", e);
-                        Err(FileListError::ModifiedTimestamp(tpath))
-                    }
-                }
-            } else {
-                Err(FileListError::FilteredOut(tpath))
-            }
-        });
-        Ok(task)
-    }).collect::<Vec<_>>();
-
-    // Here we collect the results of all tasks.
-    let mut v = Vec::new();
-    for r in tasks {
-        match r {
-            Ok(join_handle) => {
-                match join_handle.await {
-                    Ok(r) => {
-                        match r {
-                            Ok((system_time, entry)) => {
-                                v.push((system_time, entry));
-                            }
-                            Err(e) => {
-                                if let FileListError::FilteredOut(_) = e {
-                                    //
-                                } else {
-                                    error!("{:#?}", e);
-                                }
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        error!("{:#?}", e);
-                    }
-                }
-            }
-            Err(e) => {
-                if let FileListError::FilteredOut(_) = e {
-                    //
-                } else {
-                    error!("{:#?}", e);
-                }
-            }
-        }
-    }
-    debug!("Listing files in: {:#?} completed", path);
-
-    v
+fn dir_entry_path(entry: &walkdir::DirEntry) -> OsString {
+    entry
+        .path()
+        .file_name()
+        .unwrap_or_else(|| std::ffi::OsStr::new("UNKNOWN FILENAME"))
+        .to_os_string()
 }
